Refactor ROS 2 node as a component
